Update and rename lab1_var1.c to lab1_var11.c

+++ lab1/lab1_var11.c
@@ -3,13 +3,13 @@
 int main() {
     const int sizeStr = 50;
     char str[sizeStr];
-    int num;
+    float num;
     printf("Input string: ");
     fgets(str, sizeStr, stdin);
     printf("Output string: %s\n", str);
     fflush(stdin);
     printf("Input number: ");
-    scanf("%d", &num);
-    printf("Output number: %d", num);
+    scanf("%f", &num);
+    printf("Output number: %f", num);
     return 0;
 }
